requirements.txt
numpy>=1.20.0
web3>=5.30.0
requests>=2.25.0
scikit-learn>=1.0.0
pandas>=1.3.0
flask>=2.0.0
flask-cors>=3.0.10
flask-limiter>=2.7.0
ariadne>=0.18.0
pyjwt>=2.3.0
python-dotenv>=0.19.0
solana-py>=0.24.0
base58>=2.1.1
matplotlib>=3.5.0
pysha3>=1.0.2
merklelib>=1.0.1
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
eth-typing>=3.0.0
eth-utils>=2.0.0 